Complete quotes-api exercise
This is pretty crude. Obviously using an array in memory instead of a
database is not a good idea. What's also not very clear is what's the
point of having an id field and then not using that but the array index
instead to return the corresponding quote.

Many bad design decisions I think.